requirements.txt
absl-py==2.3.1
accelerate==1.12.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.3
cycler==0.12.1
diffusers==0.35.2
facenet-pytorch==2.6.0
fastapi==0.123.9
filelock==3.20.0
flatbuffers==25.9.23
fonttools==4.61.0
fsspec==2025.12.0
h11==0.16.0
hf-xet==1.2.0
httptools==0.7.1
huggingface-hub==0.36.0
idna==3.11
importlib_metadata==8.7.0
jax==0.7.1
jaxlib==0.7.1
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.7
mediapipe==0.10.21
ml_dtypes==0.5.4
mpmath==1.3.0
networkx==3.6
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opt_einsum==3.4.0
packaging==25.0
pillow==10.2.0
protobuf==4.25.8
psutil==7.1.3
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
safetensors==0.7.0
scipy==1.16.3
sentencepiece==0.2.1
setuptools==80.9.0
six==1.17.0
sounddevice==0.5.3
starlette==0.50.0
sympy==1.14.0
tokenizers==0.22.1
torch==2.2.2
torchvision==0.17.2
tqdm==4.67.1
transformers==4.57.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1
zipp==3.23.0
